test: split test_actor.py into core and TTL test files
test_actor.py: 643 → 401 lines (core actor tests)
test_actor_ttl.py: NEW 267 lines (TTL tracking and terminal notification tests)
